Conflicts:
	Gemfile
	app/assets/javascripts/jquery.lightbox-0.5.js
	app/models/p_album.rb
	app/models/upload_photo.rb
	app/models/user.rb
	app/views/comments/_form.html.erb
	app/views/upload_photos/show.html.erb

+++ app/controllers/upload_photos_controller.rb
@@ -2,13 +2,12 @@
 class UploadPhotosController < ApplicationController
   # GET /upload_photos
   # GET /upload_photos.json
-  before_filter :authenticate_user!, :except => [:index,
-                                             :show]
+  before_filter :login_required
   
   before_filter :find_photo, :only => [:show, :edit, :update, :destroy]
 
   def index
-    @upload_photos = UploadPhoto.where(id: current_user.id)
+    @upload_photos = UploadPhoto.where(user_id: current_user.id)
 
     respond_to do |format|
       format.html # index.html.erb
@@ -49,18 +48,12 @@ class UploadPhotosController < ApplicationController
         description: options[:description],
         photoURL: uploadPhoto(options[:photoURL]),
         albumBelongTo: options[:albumBelongTo],
-        user_id: options[:user_id]
+        user_id: current_user.id
     )
-      respond_to do |format|
-      # #   if @upload_photo.save
-          format.html { redirect_to @upload_photo, :notice => '照片上传成功。' }
-          format.json { render :json => @upload_photo, :status => :created, :location => @upload_photo }
-      #   # else
-      #   #   format.html { render :action => "new" }
-      #   #   format.json { render :json => @upload_photo.errors, :status => :unprocessable_entity }
-      #   # end
-      end
-    #end
+    respond_to do |format|
+      format.html { redirect_to @upload_photo, :notice => '照片上传成功。' }
+      format.json { render :json => @upload_photo, :status => :created, :location => @upload_photo }
+    end
   end
 
   # PUT /upload_photos/1
@@ -99,21 +92,18 @@ class UploadPhotosController < ApplicationController
   end
   
   def uploadPhoto(file)
-    @fileName=file.original_filename
-    if !@fileName.empty?
-      @fileName=@fileName.split(" ").to_s
+    fileName=file.original_filename
+    if !fileName.empty?
       FileUtils.mkdir($uploadFolder) unless File.exist?($uploadFolder)
-      File.open($uploadFolder+"/#{@fileName}", "wb") do |f|
+      File.open($uploadFolder+"/#{fileName}", "wb") do |f|
         f.write(file.read)
       end
-      return @fileName
+      return fileName
     end
   end
   
   def delPhoto(filename)
-    if !filename.empty?
-      FileUtils.rm($uploadFolder+"/"+filename)
-    end
+      FileUtils.rm($uploadFolder+"/"+filename) unless filename.empty?
   end
   
   def checkFileType(fileName)
